move things around and make public
